Finished Uploading products and checking browsing history.
Upload product:
When a seller uploads a product, it will save the product to the products available on the site using local storage, and saves the products to the seller's account.

Check purchase history:
When a product is bought, the purchase information will be saved to the user's account, which then will be displayed when checking the user's purchase history.